feat(#83): refactor: ask-user.ts — audit and modularize
